requirements.txt
loguru==0.5.3
pyTelegramBotAPI==4.2.2
python-dotenv==0.19.2
